@@ -1,74 +1,84 @@
 {
-    "name": "@treo/starter",
-    "version": "1.0.1",
-    "license": "https://themeforest.net/licenses/standard",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve --open",
-        "start:mem": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng serve --open",
-        "build": "ng build",
-        "build:prod": "ng build --configuration production",
-        "build:prod:mem": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "tw": "npm run tw:build && npm run tw:export",
-        "tw:build": "./node_modules/.bin/tailwind build src/tailwind/main.css -c src/tailwind/config.js -o src/styles/tailwind.scss",
-        "tw:export": "npm run tw:export:js && npm run tw:export:scss",
-        "tw:export:js": "node src/@treo/tailwind/export.js -c src/tailwind/config.js -o src/@treo/tailwind/exported/variables.ts",
-        "tw:export:scss": "./node_modules/.bin/tailwind build src/@treo/tailwind/export.css -c src/tailwind/config.js -o src/@treo/tailwind/exported/_variables.scss"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "v13-lts",
-        "@angular/cdk": "v13-lts",
-        "@angular/common": "v13-lts",
-        "@angular/compiler": "v13-lts",
-        "@angular/core": "v13-lts",
-        "@angular/forms": "v13-lts",
-        "@angular/material": "v13-lts",
-        "@angular/material-moment-adapter": "v13-lts",
-        "@angular/platform-browser": "v13-lts",
-        "@angular/platform-browser-dynamic": "v13-lts",
-        "@angular/router": "v13-lts",
-        "@types/humanize-duration": "^3.27.1",
-        "crypto-js": "^4.1.1",
-        "highlight.js": "^11.6.0",
-        "humanize-duration": "^3.27.3",
-        "lodash": "4.17.21",
-        "moment": "^2.29.4",
-        "ng-apexcharts": "^1.7.4",
-        "ngx-markdown": "^13.1.0",
-        "perfect-scrollbar": "^1.5.5",
-        "quill": "^1.3.7",
-        "rrule": "^2.7.1",
-        "rxjs": "^7.5.7",
-        "tslib": "^2.4.1",
-        "web-animations-js": "^2.3.2"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "v13-lts",
-        "@angular/cli": "v13-lts",
-        "@angular/compiler-cli": "v13-lts",
-        "@angular/language-service": "v13-lts",
-        "@types/crypto-js": "^4.1.1",
-        "@types/highlight.js": "^10.1.0",
-        "@types/jasmine": "^4.3.0",
-        "@types/jasminewd2": "^2.0.10",
-        "@types/lodash": "^4.14.188",
-        "@types/node": "^18.11.9",
-        "codelyzer": "^6.0.2",
-        "jasmine-core": "^4.5.0",
-        "jasmine-spec-reporter": "^7.0.0",
-        "karma": "^6.4.1",
-        "karma-chrome-launcher": "^3.1.1",
-        "karma-coverage": "^2.2.0",
-        "karma-jasmine": "^5.1.0",
-        "karma-jasmine-html-reporter": "^2.0.0",
-        "protractor": "^7.0.0",
-        "tailwindcss": "^3.2.3",
-        "ts-node": "^10.9.1",
-        "tslint": "^6.1.3",
-        "typescript": "^4.6.4"
-    }
+  "name": "@treo/starter",
+  "version": "1.0.1",
+  "license": "https://themeforest.net/licenses/standard",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --open",
+    "start:mem": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng serve --open",
+    "build": "ng build",
+    "build:prod": "ng build --configuration production",
+    "build:prod:mem": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "lint:fix": "ng lint --fix",
+    "format": "prettier --write \"src/**/*.{ts,html,json}\"",
+    "tw": "npm run tw:build && npm run tw:export",
+    "tw:build": "./node_modules/.bin/tailwind build src/tailwind/main.css -c src/tailwind/config.js -o src/styles/tailwind.scss",
+    "tw:export": "npm run tw:export:js && npm run tw:export:scss",
+    "tw:export:js": "node src/@treo/tailwind/export.js -c src/tailwind/config.js -o src/@treo/tailwind/exported/variables.ts",
+    "tw:export:scss": "./node_modules/.bin/tailwind build src/@treo/tailwind/export.css -c src/tailwind/config.js -o src/@treo/tailwind/exported/_variables.scss"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^21.0.5",
+    "@angular/cdk": "^16.2.14",
+    "@angular/common": "^21.0.5",
+    "@angular/compiler": "^21.0.5",
+    "@angular/core": "^21.0.5",
+    "@angular/forms": "^21.0.5",
+    "@angular/material": "^16.2.14",
+    "@angular/material-moment-adapter": "^16.2.14",
+    "@angular/platform-browser": "^21.0.5",
+    "@angular/platform-browser-dynamic": "^21.0.5",
+    "@angular/router": "^21.0.5",
+    "@types/humanize-duration": "^3.27.1",
+    "crypto-js": "^4.1.1",
+    "highlight.js": "^11.6.0",
+    "humanize-duration": "^3.27.3",
+    "lodash": "4.17.21",
+    "marked": "^17.0.1",
+    "moment": "^2.29.4",
+    "ng-apexcharts": "^1.17.1",
+    "perfect-scrollbar": "^1.5.5",
+    "quill": "^1.3.7",
+    "rrule": "^2.7.1",
+    "rxjs": "^7.5.7",
+    "tslib": "^2.4.1",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "^0.15.1"
+  },
+  "devDependencies": {
+    "@angular-eslint/builder": "~13.0.0",
+    "@angular-eslint/eslint-plugin": "~13.0.0",
+    "@angular-eslint/eslint-plugin-template": "~13.0.0",
+    "@angular-eslint/template-parser": "~13.0.0",
+    "@angular/build": "^21.0.3",
+    "@angular/cli": "^21.0.3",
+    "@angular/compiler-cli": "^21.0.5",
+    "@angular/language-service": "^21.0.5",
+    "@types/crypto-js": "^4.1.1",
+    "@types/highlight.js": "^10.1.0",
+    "@types/jasmine": "^4.3.0",
+    "@types/lodash": "~4.14.182",
+    "@types/node": "^20.19.27",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "apexcharts": "~3.35.0",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^8.10.2",
+    "eslint-plugin-prettier": "^4.2.5",
+    "jasmine-core": "^4.5.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-coverage": "^2.2.0",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "ngx-markdown": "^21.0.1",
+    "prettier": "^2.8.8",
+    "tailwindcss": "^3.2.3",
+    "ts-node": "^10.9.1",
+    "typescript": "~5.9"
+  }
 }
