@@ -11,7 +11,7 @@ module.exports = function (config)
             require('karma-chrome-launcher'),
             require('karma-jasmine-html-reporter'),
             require('karma-coverage'),
-            require('@angular-devkit/build-angular/plugins/karma')
+            
         ],
         client: {
             clearContext: false // leave Jasmine Spec Runner output visible in browser
