@@ -8,11 +8,12 @@ import { TreoNavigationService } from '@treo/components/navigation/navigation.se
 import { TreoNavigationItem } from '@treo/components/navigation/navigation.types';
 
 @Component({
-    selector       : 'treo-vertical-navigation-collapsable-item',
-    templateUrl    : './collapsable.component.html',
-    styles         : [],
-    animations     : TreoAnimations,
-    changeDetection: ChangeDetectionStrategy.OnPush
+    selector: 'treo-vertical-navigation-collapsable-item',
+    templateUrl: './collapsable.component.html',
+    styles: [],
+    animations: TreoAnimations,
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    standalone: false
 })
 export class TreoVerticalNavigationCollapsableItemComponent implements OnInit, OnDestroy
 {
