@@ -13,7 +13,8 @@ import { ScrollbarGeometry, ScrollbarPosition } from '@treo/directives/scrollbar
 // -----------------------------------------------------------------------------------------------------
 @Directive({
     selector: '[treoScrollbar]',
-    exportAs: 'treoScrollbar'
+    exportAs: 'treoScrollbar',
+    standalone: false
 })
 export class TreoScrollbarDirective implements OnInit, OnDestroy
 {
