@@ -1,5 +1,5 @@
 import {Inject, Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import {TREO_APP_CONFIG} from '@treo/services/config/config.constants';
 import {BehaviorSubject, Observable} from 'rxjs';
 import {getBasePath} from '../../app.routing';
