@@ -1,11 +1,12 @@
 import {Component, Inject, OnInit} from '@angular/core';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
+import {MAT_DIALOG_DATA, MatDialogRef as MatDialogRef} from '@angular/material/dialog';
 import {DashboardDeviceArchiveDialogService} from 'app/layout/common/dashboard-device-archive-dialog/dashboard-device-archive-dialog.service';
 
 @Component({
-  selector: 'app-dashboard-device-archive-dialog',
-  templateUrl: './dashboard-device-archive-dialog.component.html',
-  styleUrls: ['./dashboard-device-archive-dialog.component.scss'],
+    selector: 'app-dashboard-device-archive-dialog',
+    templateUrl: './dashboard-device-archive-dialog.component.html',
+    styleUrls: ['./dashboard-device-archive-dialog.component.scss'],
+    standalone: false
 })
 export class DashboardDeviceArchiveDialogComponent implements OnInit {
 
