@@ -1,10 +1,10 @@
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
-import {MatButtonModule} from '@angular/material/button';
+import {MatButtonModule as MatButtonModule} from '@angular/material/button';
 import {MatIconModule} from '@angular/material/icon';
 import {SharedModule} from 'app/shared/shared.module';
 import {dashboardRoutes} from 'app/modules/dashboard/dashboard.routing';
-import {MatDialogModule} from '@angular/material/dialog';
+import {MatDialogModule as MatDialogModule} from '@angular/material/dialog';
 import {DashboardDeviceArchiveDialogComponent} from './dashboard-device-archive-dialog.component';
 
 @NgModule({
