@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 import { DetailSettingsComponent } from './detail-settings.component';
@@ -7,7 +7,7 @@ describe('DetailSettingsComponent', () => {
   let component: DetailSettingsComponent;
   let fixture: ComponentFixture<DetailSettingsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(() => {
     TestBed.configureTestingModule({
       declarations: [ DetailSettingsComponent ],
       providers: [
@@ -15,7 +15,7 @@ describe('DetailSettingsComponent', () => {
       ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(DetailSettingsComponent);
