@@ -1,19 +1,19 @@
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
-import {MatAutocompleteModule} from '@angular/material/autocomplete';
-import {MatButtonModule} from '@angular/material/button';
-import {MatSelectModule} from '@angular/material/select';
-import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatAutocompleteModule as MatAutocompleteModule} from '@angular/material/autocomplete';
+import {MatButtonModule as MatButtonModule} from '@angular/material/button';
+import {MatSelectModule as MatSelectModule} from '@angular/material/select';
+import {MatFormFieldModule as MatFormFieldModule} from '@angular/material/form-field';
 import {MatIconModule} from '@angular/material/icon';
-import {MatInputModule} from '@angular/material/input';
+import {MatInputModule as MatInputModule} from '@angular/material/input';
 import {SharedModule} from 'app/shared/shared.module';
 import {DetailSettingsComponent} from 'app/layout/common/detail-settings/detail-settings.component'
-import {MatDialogModule} from '@angular/material/dialog';
+import {MatDialogModule as MatDialogModule} from '@angular/material/dialog';
 import {MatButtonToggleModule} from '@angular/material/button-toggle';
-import {MatTabsModule} from '@angular/material/tabs';
-import {MatSliderModule} from '@angular/material/slider';
-import {MatSlideToggleModule} from '@angular/material/slide-toggle';
-import {MatTooltipModule} from '@angular/material/tooltip';
+import {MatTabsModule as MatTabsModule} from '@angular/material/tabs';
+import {MatSliderModule as MatSliderModule} from '@angular/material/slider';
+import {MatSlideToggleModule as MatSlideToggleModule} from '@angular/material/slide-toggle';
+import {MatTooltipModule as MatTooltipModule} from '@angular/material/tooltip';
 
 @NgModule({
     declarations: [
