@@ -7,10 +7,11 @@ import { TreoNavigationService } from '@treo/components/navigation';
 import {versionInfo} from 'environments/versions';
 
 @Component({
-    selector     : 'material-layout',
-    templateUrl  : './material.component.html',
-    styleUrls    : ['./material.component.scss'],
-    encapsulation: ViewEncapsulation.None
+    selector: 'material-layout',
+    templateUrl: './material.component.html',
+    styleUrls: ['./material.component.scss'],
+    encapsulation: ViewEncapsulation.None,
+    standalone: false
 })
 export class MaterialLayoutComponent implements OnInit, OnDestroy
 {
